REFACTOR: Namings
