+++ Neos.Neos/Tests/Behavior/Features/Bootstrap/RoutingTrait.php
@@ -46,8 +46,8 @@ use Neos\Neos\Domain\Model\Site;
 use Neos\Neos\Domain\Model\SiteIdentifier;
 use Neos\Neos\Domain\Repository\DomainRepository;
 use Neos\Neos\Domain\Repository\SiteRepository;
-use Neos\Neos\EventSourcedRouting\NodeUriBuilder;
-use Neos\Neos\EventSourcedRouting\Projection\DocumentUriPathProjector;
+use Neos\Neos\FrontendRouting\NodeUriBuilder;
+use Neos\Neos\FrontendRouting\Projection\DocumentUriPathProjector;
 use Neos\Neos\FrontendRouting\DimensionResolution\RequestToDimensionSpacePointContext;
 use Neos\Neos\FrontendRouting\DimensionResolution\DimensionResolverFactoryInterface;
 use Neos\Neos\FrontendRouting\EventSourcedFrontendNodeRoutePartHandler;
